// frontend/src/app/model/user.ts
export class User {
    id: string;
    roleId: string;
    userAccountId: string;
    firstName: string;
    lastName: string;
    address: string;
    age: string;
    email: string;
    password: string;
    phone: string;
    country: string;
    insertDateTime: Date;
    updateDateTime: Date;
    status: number;
    sortOrder: number;
    averageRating: number;
    update: string;
}